Changed some ints to size_ts to eliminate annoying compiler warnings.

+++ src/stan/io/reader.hpp
@@ -979,21 +979,21 @@ namespace stan {
       template <typename TL>
       inline vector_t vector_lb(const TL lb, size_t m) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lb(lb);
         return v;
       }
       template <typename TL>
       inline vector_t vector_lb_constrain(const TL lb, size_t m) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lb_constrain(lb);
         return v;
       }
       template <typename TL>
       inline vector_t vector_lb_constrain(const TL lb, size_t m, T& lp) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lb_constrain(lb,lp);
         return v;
       }
@@ -1001,21 +1001,21 @@ namespace stan {
       template <typename TL>
       inline row_vector_t row_vector_lb(const TL lb, size_t m) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lb(lb);
         return v;
       }
       template <typename TL>
       inline row_vector_t row_vector_lb_constrain(const TL lb, size_t m) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lb_constrain(lb);
         return v;
       }
       template <typename TL>
       inline row_vector_t row_vector_lb_constrain(const TL lb, size_t m, T& lp) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lb_constrain(lb,lp);
         return v;
       }
@@ -1023,24 +1023,24 @@ namespace stan {
       template <typename TL>
       inline matrix_t matrix_lb(const TL lb, size_t m, size_t n) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j  = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j  = 0; j < n; ++j)
             v(i,j) = scalar_lb(lb);
         return v;
       }
       template <typename TL>
       inline matrix_t matrix_lb_constrain(const TL lb, size_t m, size_t n) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j = 0; j < n; ++j)
             v(i,j) = scalar_lb_constrain(lb);
         return v;
       }
       template <typename TL>
       inline matrix_t matrix_lb_constrain(const TL lb, size_t m, size_t n, T& lp) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j = 0; j < n; ++j)
             v(i,j) = scalar_lb_constrain(lb,lp);
         return v;
       }
@@ -1050,21 +1050,21 @@ namespace stan {
       template <typename TU>
       inline vector_t vector_ub(const TU ub, size_t m) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_ub(ub);
         return v;
       }
       template <typename TU>
       inline vector_t vector_ub_constrain(const TU ub, size_t m) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_ub_constrain(ub);
         return v;
       }
       template <typename TU>
       inline vector_t vector_ub_constrain(const TU ub, size_t m, T& lp) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_ub_constrain(ub,lp);
         return v;
       }
@@ -1072,21 +1072,21 @@ namespace stan {
       template <typename TU>
       inline row_vector_t row_vector_ub(const TU ub, size_t m) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_ub(ub);
         return v;
       }
       template <typename TU>
       inline row_vector_t row_vector_ub_constrain(const TU ub, size_t m) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_ub_constrain(ub);
         return v;
       }
       template <typename TU>
       inline row_vector_t row_vector_ub_constrain(const TU ub, size_t m, T& lp) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_ub_constrain(ub,lp);
         return v;
       }
@@ -1094,24 +1094,24 @@ namespace stan {
       template <typename TU>
       inline matrix_t matrix_ub(const TU ub, size_t m, size_t n) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j  = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j  = 0; j < n; ++j)
             v(i,j) = scalar_ub(ub);
         return v;
       }
       template <typename TU>
       inline matrix_t matrix_ub_constrain(const TU ub, size_t m, size_t n) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j = 0; j < n; ++j)
             v(i,j) = scalar_ub_constrain(ub);
         return v;
       }
       template <typename TU>
       inline matrix_t matrix_ub_constrain(const TU ub, size_t m, size_t n, T& lp) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j = 0; j < n; ++j)
             v(i,j) = scalar_ub_constrain(ub,lp);
         return v;
       }
@@ -1120,21 +1120,21 @@ namespace stan {
       template <typename TL, typename TU>
       inline vector_t vector_lub(const TL lb, const TU ub, size_t m) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lub(lb,ub);
         return v;
       }
       template <typename TL, typename TU>
       inline vector_t vector_lub_constrain(const TL lb, const TU ub, size_t m) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lub_constrain(lb,ub);
         return v;
       }
       template <typename TL, typename TU>
       inline vector_t vector_lub_constrain(const TL lb, const TU ub, size_t m, T& lp) {
         vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lub_constrain(lb,ub,lp);
         return v;
       }
@@ -1142,21 +1142,21 @@ namespace stan {
       template <typename TL, typename TU>
       inline row_vector_t row_vector_lub(const TL lb, const TU ub, size_t m) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lub(lb,ub);
         return v;
       }
       template <typename TL, typename TU>
       inline row_vector_t row_vector_lub_constrain(const TL lb, const TU ub, size_t m) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lub_constrain(lb,ub);
         return v;
       }
       template <typename TL, typename TU>
       inline row_vector_t row_vector_lub_constrain(const TL lb, const TU ub, size_t m, T& lp) {
         row_vector_t v(m);
-        for (int i = 0; i < m; ++i)
+        for (size_t i = 0; i < m; ++i)
           v(i) = scalar_lub_constrain(lb,ub,lp);
         return v;
       }
@@ -1164,24 +1164,24 @@ namespace stan {
       template <typename TL, typename TU>
       inline matrix_t matrix_lub(const TL lb, const TU ub, size_t m, size_t n) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j  = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j  = 0; j < n; ++j)
             v(i,j) = scalar_lub(lb,ub);
         return v;
       }
       template <typename TL, typename TU>
       inline matrix_t matrix_lub_constrain(const TL lb, const TU ub, size_t m, size_t n) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j = 0; j < n; ++j)
             v(i,j) = scalar_lub_constrain(lb,ub);
         return v;
       }
       template <typename TL, typename TU>
       inline matrix_t matrix_lub_constrain(const TL lb, const TU ub, size_t m, size_t n, T& lp) {
         matrix_t v(m,n);
-        for (int i = 0; i < m; ++i)
-          for (int j = 0; j < n; ++j)
+        for (size_t i = 0; i < m; ++i)
+          for (size_t j = 0; j < n; ++j)
             v(i,j) = scalar_lub_constrain(lb,ub,lp);
         return v;
       }
